Add loading-state component to PostArticle
This means we can move isLoading? behavior out of a parent
component and simply pass isLoading down to the PostArticle which
is responsible for the presentation of its own loading-state.

Currently it shows gray-box type placeholders.

PostArticle remains stateless and relies on parents to fetch
and supply post data/content.