Add LoginChallenge, Modal, useUser
This commit adds a LoginChallenge component to the draft posts page
and to the edit post page.
LoginChallenge presents as a model with the Login form, if the user
is not logged in.

The checkLogin function has been replaced with the useUser hook as
recommended in the useSWR docs.

The Modal component is not fullly accessibile yet. And this whole
thing probably makes a lot more sense if we use context to keep
track of the user.