Upload images and return their new URL
This adds an ImageForm component which currently only
makes new images, and isn't very graceful with handling
isDirty and isSubmitting and so on (with some UX touches
missing and other jerry-rigged). It should probably have
a "reset" or "clear" button, and then it should make up
its mind about whether it is also handling "delete", and
"replace" type actions.