Move index page to /app router, recreate layout
This meant splitting the Layout component into a layout.js RootLayout
plus a (public) and (private) folder with their own layout files. The
result I think is quite nice and less complex than before. Instead of
passing props from the page to the layout, and having lots of if/then
(or ternaries to set Tailwind classes), we just have the root layout,
and then for the public layout we add the banner and footer (removing
the if/then for this). Then we moved the "singleCol" treatment to the
private layout, but later we'll probably move this too.

This commit also creates an IffLoggedIn which is a client component
that can wrap other components that should only be displayed when
logged in. The previous approach required using the useSession hook
and displaying the contents if/then-style, which requires the entire
component to run on the client! Now a server component or statically
rendered page can include login-only content by wrapping it in:

```jsx
<IffLoggedIn> e.g. display 'edit post' button </IffLoggedIn>
```