Add cyan and lilac custom theme colors
This commit also ditches the -100 -200 ... -950 scale for the cyan colors,
and adds a lilac color, and replaces the color-scales with semantic names:

* cyan, cyan-soft, cyan-content, cyan-bright
* lilac, lilac-soft, lilac-content, lilac-bright

The profile pic backrop is `lilac` which gets saturated to make `lilac-bright`
and the tshirt/eyes color is `cyan-bright`, which gets de-saturated to make
`cyan`. `cyan` and `lilac` base colors are then lightened or darkened to make
the `-soft` and `-content` variants.